Adding more code
Game engine is, tragically, not done yet. However, a lot of the real
meat of the classes has been laid out in Position, which basically
everything inherits from to some extent. I'll work on this a little more
in Period 1, specifically polishing up Position and trying to finish
Engine. The rest is simple and boilerplate-y, meaning it's a bit
repetitive and something you guys can help me with :smile: